+++ client/src/components/main/BackgroundCanvas.tsx
