+++ client/src/components/main/BackgroundImage.tsx
